product variant item feture

+++ app/Http/Controllers/Backend/ProductVariantController.php
@@ -6,6 +6,7 @@ use App\DataTables\ProductVariantDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\Product;
 use App\Models\ProductVariant;
+use App\Models\ProductVariantItem;
 use Illuminate\Http\Request;
 
 class ProductVariantController extends Controller
@@ -92,6 +93,11 @@ class ProductVariantController extends Controller
     public function destroy(string $id)
     {
         $variant = ProductVariant::findOrFail($id);
+        $variantItemcheck = ProductVariantItem::where('product_variant_id', $variant->id)->count();
+        if ($variantItemcheck > 0) {
+            return response(['status' => 'error', 'message' => 'This variant containt variant in it delete
+            the variant item first for delete this variant!']);
+        }
         $variant->delete();
 
         return response(['status' => 'success', 'message' => 'Deleted Successfully!']);
